eatgo-api/src/test/java/kr/ardent/eatgo/application/ReviewServiceTests.java
package kr.ardent.eatgo.application;

import kr.ardent.eatgo.domain.Review;
import kr.ardent.eatgo.domain.ReviewRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;

import static org.junit.Assert.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;

public class ReviewServiceTests {

    private ReviewService reviewService;
    @Mock
    private ReviewRepository reviewRepository;

    @Before
    public void setUp(){
        MockitoAnnotations.initMocks(this);
        reviewService = new ReviewService(reviewRepository);
    }

    @Test
    public void review(){
        Review review = Review.builder().name("JJ").score(5).description("Good!").build();
        reviewService.addReview(review);

        verify(reviewRepository).save(any());
    }

}